requirements.txt
python-dotenv
crewai
crewai-tools
langchain-openai